requirements.txt
opencv_python==4.5.5.64
Pillow==9.2.0
pytest==7.1.2
requests==2.28.1
setuptools==57.4.0
youtube_dl==2021.6.6
